preserving hilite_re in parser/wiki.py
Observed that the highlighting was not working when using the
Check Spelling function... this made it somewhat useless.  Tracing
through found that wiki.py was hard-coding hilite_re to ''.
Added a conditional to set it to self.formatter.page.hilite_re if it
exists; if not, it sets hilite_re to '' to avoid breaking things
downstream.  Hopefully.

Thanks to Brad Mandell for finding the bug!